pyproject.toml
[tool.poetry]
name = "salute"
package-mode = false
authors = ["Dan Trickey <git@trickey.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
django-choices-field = "^2.3.0"
django-cors-headers = "^4.7.0"
django-environ = "^0.12.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.1.0"}
google-api-python-client = "^2.174.0"
joserfc = "^1.1.0"
pydantic = "^2.11.7"
requests = "^2.32.4"
rules = "^3.5"
strawberry-graphql = "~0.275.5"
strawberry-graphql-django = "~0.60.0"  # Issue in 0.61.0 to do with lazy ordering type resolution
tqdm = "^4.67.1"
whitenoise = "^6.9.0"
django-birdbath = "^2.0.1"
xkcdpass = "^1.20.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.2.0"
django-stubs = {version = "^5.2.1", extras = ["compatible-mypy"]}
factory-boy = "3.3.2"  # 3.3.3 exports incompatible type hints
mypy = "^1.16.1"
pytest = "^8.4.1"
ruff = "^0.12.1"
psycopg = {extras = ["binary"], version = "^3.2.9"}
pytest-cov = "^6.2.1"
pytest-django = "^4.11.1"
types-requests = "^2.32.4.20250611"
responses = "^0.25.7"
types-tqdm = "^4.67.0.20250516"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.14"

[tool.poetry.group.prod.dependencies]
granian = "^2.3.4"
psycopg = "^3.2.9"
sentry-sdk = {extras = ["django"], version = "^2.32.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "A",
    "ANN",
    "B",
    "BLE",
    "C4",
    "COM",
    "DJ",
    "DTZ",
    "E",
    "F",
    "FA",
    "FBT",
    "I",
    "N",
    "S",
    "T10",
    "UP",
    "W",
]
ignore = [
    "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed
    "B009",    # Do not call `getattr` with a constant attribute value.
    "COM812",  # Conflicts with formatter
    "N999",    # Invalid module name
    "S101",    # S101 Use of `assert` detected
]

[tool.ruff.lint.isort.sections]
# Group all Django imports into a separate section.
"django" = ["django"]

[tool.ruff.lint.per-file-ignores]
"*/birdbath.py" = ["ANN001", "ANN201"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

warn_unused_ignores = true

strict_optional = true
no_implicit_optional = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["debug_toolbar.*", "factory.*", "googleapiclient.discovery.*", "phonenumber_field.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["*.tests.*"]
disable_error_code = ["arg-type"]

[[tool.mypy.overrides]]
module = ["*.birdbath"]
ignore_errors = true

[tool.django-stubs]
django_settings_module = "salute.settings.dev"

[tool.coverage.run]
omit = [
    # omit migrations and tests from test coverage
    "*/migrations/*",
    "*/tests/*",
]